Deploy via docker

+++ app/services/product_service.py
@@ -1,7 +1,7 @@
 from sqlalchemy.orm import Session
 from app.models.product_model import Product
 from app.schemas.product_schema import ProductCreate, ProductUpdate
-
+from app.models.order_model import OrderItem
 def create_product(db: Session, product_data: ProductCreate):
     new_product = Product(**product_data.dict())
     db.add(new_product)
@@ -26,7 +26,12 @@ def update_product(db: Session, product_id: int, data: ProductUpdate):
     return product
 
 def delete_product(db: Session, product_id: int):
+    count = db.query(OrderItem).filter(OrderItem.product_id == product_id).count()
+    if count > 0:
+        return False
     product = get_product_by_id(db, product_id)
     if product:
         db.delete(product)
         db.commit()
+        return True
+    return False
